refactor: change project and templates layout
